refactor: Making `Menu` component reusable
